+++ NEWS.md
@@ -1,5 +1,7 @@
 # version 1.0-22
 
+* handle empty points in `st_distance()`; https://github.com/r-spatial/s2/issues/289
+
 * for unprojected lines, `st_line_interpolate()` requires distance values with degree units; #2542
 
 * `unique.sfc()` added; #2546